Make the new function wrapper work.
